+++ src/Functions/Set-AppSettings-For-Function-App.ps1
@@ -4,6 +4,7 @@ param (
     [Parameter(Mandatory)][string] $FunctionAppName,
     [Parameter(Mandatory)][string[]] $FunctionAppAppSettings,
     [Parameter()][string] $FunctionAppDeploymentSlotName,
+    [Parameter()][bool] $EnableDeploymentSetting = $false,
     [Parameter()][bool] $ApplyToAllSlots = $false
 )
 
@@ -13,22 +14,28 @@ Import-Module "$PSScriptRoot\..\AzDocs.Common" -Force
 
 Write-Header -ScopedPSCmdlet $PSCmdlet
 
-if ($ApplyToAllSlots)
-{
+if ($ApplyToAllSlots) {
     $availableSlots = Invoke-Executable -AllowToFail az functionapp deployment slot list --name $FunctionAppName --resource-group $FunctionAppResourceGroupName | ConvertFrom-Json
 }
 
 $optionalParameters = @()
-if ($FunctionAppDeploymentSlotName)
-{
+if ($FunctionAppDeploymentSlotName) {
     $optionalParameters += "--slot", "$FunctionAppDeploymentSlotName"
 }
 
-Invoke-Executable az functionapp config appsettings set --resource-group $FunctionAppResourceGroupName --name $FunctionAppName --settings @FunctionAppAppSettings @optionalParameters
+if ($EnableDeploymentSetting) {
+    Invoke-Executable az functionapp config appsettings set --resource-group $FunctionAppResourceGroupName --name $FunctionAppName --slot-settings @FunctionAppAppSettings @optionalParameters
 
-foreach($availableSlot in $availableSlots)
-{
-    Invoke-Executable az functionapp config appsettings set --resource-group $FunctionAppResourceGroupName --name $FunctionAppName --settings @FunctionAppAppSettings --slot $availableSlot.name
+    foreach ($availableSlot in $availableSlots) {
+        Invoke-Executable az functionapp config appsettings set --resource-group $FunctionAppResourceGroupName --name $FunctionAppName --slot-settings @FunctionAppAppSettings --slot $availableSlot.name
+    }
+}
+else {
+    Invoke-Executable az functionapp config appsettings set --resource-group $FunctionAppResourceGroupName --name $FunctionAppName --settings @FunctionAppAppSettings @optionalParameters
+
+    foreach ($availableSlot in $availableSlots) {
+        Invoke-Executable az functionapp config appsettings set --resource-group $FunctionAppResourceGroupName --name $FunctionAppName --settings @FunctionAppAppSettings --slot $availableSlot.name
+    }
 }
 
 Write-Footer -ScopedPSCmdlet $PSCmdlet